Move jmx test engine into test code with exec plugin profile to execute
