+++ core/pom.xml
@@ -30,7 +30,7 @@
 
     <artifactId>dozer</artifactId>
     <packaging>bundle</packaging>
-    <name>Dozer</name>
+    <name>Dozer :: Core</name>
 
     <dependencies>
         <!-- Commons -->
@@ -269,5 +269,30 @@
             </plugin>
         </plugins>
     </build>
+
+    <profiles>
+        <profile>
+            <id>jmxtest</id>
+            <build>
+                <plugins>
+                    <plugin>
+                        <groupId>org.codehaus.mojo</groupId>
+                        <artifactId>exec-maven-plugin</artifactId>
+                        <executions>
+                            <execution>
+                                <goals>
+                                    <goal>java</goal>
+                                </goals>
+                            </execution>
+                        </executions>
+                        <configuration>
+                            <mainClass>org.dozer.jmxtest.JMXTestEngine</mainClass>
+                            <classpathScope>test</classpathScope>
+                        </configuration>
+                    </plugin>
+                </plugins>
+            </build>
+        </profile>
+    </profiles>
 
 </project>
