--- core/src/misc/docs/HowToRunJMXTestEngine.txt
@@ -1,21 +1,0 @@
-====
-    Copyright 2005-2017 Dozer Project
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-         http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-====
-
-  1.  Add misc/jmx-testengine/JMXTestEngine to your dozer eclipse project src.  Doesn't matter where.  The java file doesnt have a package so it just needs to be in root
-  2.  Using jdk1.5, run JMXTestEngine with the following VM arg set --> -Dcom.sun.management.jmxremote
-  3.  Open a dos window and run jconsole.  i.e) C:\Sun\AppServer\jdk\bin>jconsole
-  4.  When jconsole opens, select the JMXTestEngine process and click connect.  Go to the mbeans tab and you should see the 2 dozer mbeans
-
