+++ plugins-parent/pom.xml
@@ -30,7 +30,7 @@
 
     <artifactId>dozer-plugins-parent</artifactId>
     <packaging>pom</packaging>
-    <name>Dozer Plugins Parent</name>
+    <name>Dozer :: Plugins Parent</name>
 
     <properties>
         <bom-dependencies-version>6.0.0-SNAPSHOT</bom-dependencies-version>
@@ -54,6 +54,7 @@
         <!-- Codehaus Plugins versions; alphabetical order -->
         <build-helper-maven-plugin.version>3.0.0</build-helper-maven-plugin.version>
         <cobertura-maven-plugin.version>2.7</cobertura-maven-plugin.version>
+        <exec-maven-plugin.version>1.6.0</exec-maven-plugin.version>
         <findbugs-maven-plugin.version>3.0.4</findbugs-maven-plugin.version>
         <jaxb2-maven-plugin.version>2.3.1</jaxb2-maven-plugin.version>
         <versions-maven-plugin.version>2.3</versions-maven-plugin.version>
@@ -201,6 +202,11 @@
                     <groupId>org.codehaus.mojo</groupId>
                     <artifactId>cobertura-maven-plugin</artifactId>
                     <version>${cobertura-maven-plugin.version}</version>
+                </plugin>
+                <plugin>
+                    <groupId>org.codehaus.mojo</groupId>
+                    <artifactId>exec-maven-plugin</artifactId>
+                    <version>${exec-maven-plugin.version}</version>
                 </plugin>
                 <plugin>
                     <groupId>org.codehaus.mojo</groupId>
