+++ pom.xml
@@ -26,7 +26,7 @@
     <packaging>pom</packaging>
     <version>6.0.0-SNAPSHOT</version>
     <url>http://dozer.sourceforge.net</url>
-    <name>Dozer Parent Project</name>
+    <name>Dozer :: Parent</name>
     <description>Dozer is a powerful Java Bean to Java Bean mapper that recursively copies data from one object to another</description>
     <inceptionYear>2005</inceptionYear>
